CDAP-50 Base implementation to defer program starting. Now program starting has two phases: preparing and starting. So now lifecycle looks like: 1. Prepare program start and obtain controller by ProgramRuntimeService.prepare() 2. Set all listeners into controller 3. Start program from ProgramController.start()
I moved code part in runners where we start program into Callable function, pass this function to controller and this function will be triggered when we start program by ProgramController.start().

Attention! This is not complete implementation and not compilable, I have done only with WorkflowProgram, SparkProgram, ServiceProgram, WebappProgram so you can check how it looks like and if this solution will be satisfactory, I'll remade another programs.